backend/requirements.txt
torch
opencv-python
mediapipe
ultralytics
numpy
torchvision
pillow
pymysql
flask